+++ lib/printavo/models/line_item_group.rb
@@ -1,0 +1,10 @@
+# lib/printavo/models/line_item_group.rb
+# frozen_string_literal: true
+
+module Printavo
+  class LineItemGroup < Models::Base
+    def id          = self['id']
+    def name        = self['name']
+    def description = self['description']
+  end
+end