+++ lib/printavo/resources/contacts.rb
@@ -44,14 +44,6 @@ module Printavo
                                variables: { id: id.to_s, input: camelize_keys(input) })
         Printavo::Contact.new(data['contactUpdate'])
       end
-
-      private
-
-      def camelize_keys(hash)
-        hash.transform_keys do |key|
-          key.to_s.gsub(/_([a-z])/) { ::Regexp.last_match(1).upcase }
-        end
-      end
     end
   end
 end