+++ lib/printavo/resources/customers.rb
@@ -76,13 +76,6 @@ module Printavo
         end
         Printavo::Customer.new(normalized)
       end
-
-      # Converts snake_case Ruby keyword args to camelCase for GraphQL input.
-      def camelize_keys(hash)
-        hash.transform_keys do |key|
-          key.to_s.gsub(/_([a-z])/) { ::Regexp.last_match(1).upcase }
-        end
-      end
     end
   end
 end