+++ lib/printavo/resources/transaction_payments.rb
@@ -24,14 +24,6 @@ module Printavo
         @graphql.mutate(DELETE_MUTATION, variables: { id: id.to_s })
         nil
       end
-
-      private
-
-      def camelize_keys(hash)
-        hash.transform_keys do |key|
-          key.to_s.gsub(/_([a-z])/) { ::Regexp.last_match(1).upcase }
-        end
-      end
     end
   end
 end